% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssPrincomp.R
\name{biplot.dssPrincomp}
\alias{biplot.dssPrincomp}
\title{Biplot a dssPrincomp object}
\usage{
\method{biplot}{dssPrincomp}(
  x,
  choices = 1L:2L,
  type = "combine",
  levels = NULL,
  draw.arrows = TRUE,
  ...,
  datasources = NULL
)
}
\arguments{
\item{x}{an object of class dssPrincomp.}

\item{choices}{length 2 vector specifying the components to plot (same as for biplot.princomp)}

\item{type}{a character which represents the type of analysis to carry out.
If type is set to 'combine',  global column means are calculated if type is set to 'split', the column means are
calculated separately for each node.}

\item{levels}{a character the name of a factor. If provided, the plot will be colour coded by the levels of this factor}

\item{draw.arrows}{a logical, should I draw arrows representing the underlying variables? Default TRUE.}

\item{...}{further arguments to be passed to dssSmooth2d (see doc)}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
Biplot implementation for PCA on distributed datasets
}
