% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{.add.alpha}
\alias{.add.alpha}
\title{Add opacity (alpha) to a vector of colours}
\usage{
.add.alpha(old.colours, alpha)
}
\arguments{
\item{old.colours}{colour vector}

\item{alpha}{opacity ratio (1 - perfectly opaque, 0- perfectly transparent)}
}
\value{
the new colour vector with alpha
}
\description{
Add opacity (alpha) to a vector of colours
}
