% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{.encode.arg}
\alias{.encode.arg}
\title{Encode function  arguments}
\usage{
.encode.arg(some.object, serialize.it = FALSE)
}
\arguments{
\item{some.object}{the object to be encoded}
}
\value{
encoded text with offending characters replaced by strings
}
\description{
Serialize to JSON, then encode base64,
 then replace '+', '/' and '=' in the result in order to play nicely with the opal sentry.
 Used to encode non-scalar function arguments prior to sending to the opal server.
 There's a corresponding function in the server package calle .decode_args
}
