% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssLM.R
\name{.linreg.one}
\alias{.linreg.one}
\title{Use the quantities (X'X, X'y, y'y, sample size) sent from the individual datasources
  to reconstruct a complete linear model.}
\usage{
.linreg.one(XX, Xy, yy, n)
}
\description{
Use the quantities (X'X, X'y, y'y, sample size) sent from the individual datasources
  to reconstruct a complete linear model.
}
