% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssLM.R
\name{.linreg.reduce}
\alias{.linreg.reduce}
\title{Aggregate linear models into a single one for the estimation of beta.}
\usage{
.linreg.reduce(reslist)
}
\arguments{
\item{reslist:}{a list of the results from multiple calls to `linregDSS`
with the same dependent and explanatory variables.}
}
\value{
the MLE estimator for beta for the pooled individuals.
}
\description{
Takes as input multiple results from `linregDSS` to make a single linear model,
   as if individuals from the multiple sources were pooled together before modeling.
}
