% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssRandomForest.R
\name{.quality}
\alias{.quality}
\title{Measure the quality of a randomForest object.}
\usage{
.quality(forestObject)
}
\description{
The quality is the average OOB errors in the case of classification,
  or the average pseudo R-squared for regression.
  We do it in the client so that the server method can return a pure
  randomForest object (we avoid using a custom object between the two).
}
\details{
#deprecated
}
