% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{.rescale}
\alias{.rescale}
\title{Rescale a vector given a new range}
\usage{
.rescale(x, new.range, old.range = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{new.range}{a vector containing the new min and max}

\item{old.range}{(optional) the old range}
}
\value{
the rescaled vector
}
\description{
Rescale a vector given a new range
}
