% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssAddDaysInterval.R
\name{dssAddDaysInterval}
\alias{dssAddDaysInterval}
\title{Add and interval column to a dataframe}
\usage{
dssAddDaysInterval(
  newobj,
  df,
  description.list,
  async = TRUE,
  wait = TRUE,
  datasources = NULL
)
}
\arguments{
\item{newobj}{a character, name of the new dataframe}

\item{df}{a character, name of the original dataframe}

\item{description.list}{a named list containing the name of the new colunn pointing to a list containing the start date and end date.
The format is list(<new column name> = list(start_date = '<existing date column>', end_date = '<existing date column>')) - see example.}

\item{async}{same as in datashield.assign}

\item{wait}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\description{
Add a column containing the difference between 2 date columns. The date columns must be in the same dataframe
}
\examples{
# login:
opals <- datashield.login(logindata)
# load DM in colaus:
datashield.assign(opals['colaus'], 'dm', 'rhapsody.DM')
# add an interval column as the difference in days between RFSDTC and BRTHDTC:
ds2.add.days.interval('dm2', 'dm', list(DM_ELAPSED = list(end_date = 'RFSTDTC', start_date = 'BRTHDTC')), datasources = opals['colaus'])
# check the newly created column:
ds.summary('dm2$DM_ELAPSED', datasources = opals['colaus'])



}
