% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssAs.R
\name{dssAs}
\alias{dssAs}
\title{Coerce to new type (like R as... functions)}
\usage{
dssAs(
  newtype,
  object,
  newobject = NULL,
  async = TRUE,
  wait = TRUE,
  datasources = NULL
)
}
\arguments{
\item{newtype}{a character, what type to convert to to (ex: "character" or "data.frame")}

\item{object}{a character, name of the object to convert}

\item{newobject}{a character, optional, name of the newly created object. If missing the old object will be overwritten.}

\item{async}{a logical, see datashield.aggregate}

\item{wait}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\value{
TRUE if succesful
}
\description{
Coerce to new type (like R as... functions)
}
\examples{
dssAs('data.frame', 'some_matrix_object', 'my_new_df',  datsources = opals)
}
