% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssConsensusOPLS.R
\name{dssConsensusOPLS}
\alias{dssConsensusOPLS}
\title{Call the ConsensusOPLS function from the package ConsensusOPLS}
\usage{
dssConsensusOPLS(async, datasources, ...)
}
\arguments{
\item{async}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}

\item{...}{further *named* arguments to be passed to the function (see the documentation of the ConsensusOPLS package). The names of objects (dataframes, vectors)
should be provided as a characters.}
}
\description{
Call the ConsensusOPLS function from the package ConsensusOPLS
}
