% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{dssCreateFakeServers}
\alias{dssCreateFakeServers}
\title{Fake opal servers in R}
\usage{
dssCreateFakeServers(opal_name, servers = 1, tie_first_to_GlobalEnv = FALSE)
}
\arguments{
\item{opal_name}{required, a character. The name of a new list (created by the function as a side effect)
containing the pseudo servers.}

\item{servers}{either the number of servers or a vector containing their names}

\item{tie_first_to_GlobalEnv}{a logical, should the first server session be the same as .GlobalEnv? See details.}
}
\value{
a vector containing the server names. This vector (or a subset) will be used by datashield.login as first parameter.
}
\description{
Create pseudo opal/datashield servers in the local session for fun and profit.
}
\details{
The function creates a list object containing the local pseudo servers as elements. The names of the servers
are either provided in the "servers" parameter or created as 'local1', 'local2' etc. Each "server" is an environment.
If tie_first_to_GlobalEnv is set to TRUE, the first server in the list will be a reference to the global environment.
This means that all the objects in .GlobalEnv will become available to datashield... methods.
Lastly, the function overrides datashield.login from the package opal.
}
\seealso{
\code{\link{datashield.login}}
}
