% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssCut.R
\name{dssCut}
\alias{dssCut}
\title{Convert numeric to factor}
\usage{
dssCut(
  x,
  new.name = "newObj",
  df = NULL,
  in.place = FALSE,
  breaks = NULL,
  labels = NULL,
  ...,
  async = TRUE,
  wait = TRUE,
  datasources = NULL
)
}
\arguments{
\item{x}{a character name of  a numeric vector which is to be converted to a factor by cutting}

\item{new.name}{a character name of the new vector (or dataframe column). This parameter is ignored
if in.place = TRUE (see below)}

\item{df}{a character name of a dataframe. If specified, x and new.name above must be among its columns}

\item{in.place}{a boolean. If true, the numeric vector will be replaced with the output of the function}

\item{breaks}{either a numeric vector of two or more unique cut points or a single number
(greater than or equal to 2) giving the number of intervals into which x is to be cut.
Use the vector form (2 or more specific cut points) in order to achieve consistent levels among
multiple nodes.}

\item{labels}{same as for the 'cut' function. If only labels and no breaks are specified, breaks will be
calculated as length(labels)}

\item{...}{further parameters are sent directly to the cut function on each node}

\item{async}{a logical, see datashield.aggregate}

\item{wait}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\value{
name of the new object
}
\description{
Same  as base::cut
}
