% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssDDRTree.R
\name{dssDDRTree}
\alias{dssDDRTree}
\title{Call selected functions from the package DDRTree}
\usage{
dssDDRTree(func, ..., async = TRUE, datasources = NULL)
}
\arguments{
\item{func}{a character, the name of the function to call}

\item{...}{further arguments to be passed to the function (see the documentation of the DDRTree package).}

\item{async}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
At the moment only the function 'DDRTree' is implemented. Due to implementation specifics the 'X' argument
for this function is mandatory (and is specified as a character, the name of a data frame).
}
