% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssDeriveColumn.R
\name{dssDeriveColumn}
\alias{dssDeriveColumn}
\title{Add a calculated column to a dataframe}
\usage{
dssDeriveColumn(
  df,
  col.name,
  formula,
  datasources = NULL,
  async = TRUE,
  wait = TRUE
)
}
\arguments{
\item{df}{a character, the name of the dataframe}

\item{formula}{a character, the formula used to calculate the new column. It can reference columns from the dataframe, or other objects in the session}

\item{datasources}{same as in datashield.assign}

\item{async}{same as in datashield.assign}

\item{wait}{same as in datashield.assign}

\item{new.col}{a character, the name of the new column}
}
\value{
It doesn't return anything, the dataframe in the server session will now contain the new column.
}
\description{
The new column is calculated from a formula that can contain arithmetic operators and a (very) limited number of R functions.
At the moment the permitted functions are: abs, round, ceiling, floor, trunc, signif, length.
}
\examples{
# open a local pseudo connection:
opals <- dssCreatePseudoServers(servers = 1, tie_first_to_GlobalEnv = TRUE)
data('iris')
#now play around:
dssDeriveColumn('iris', 'new_col', 'Sepal.Length/round(2.2)', datasources = opals)
str(iris)
dssDeriveColumn('iris', 'new_col2', 'iris$Sepal.Length/abs(-2)', datasources = opals)
str(iris)
a <- iris$Petal.Length
dssDeriveColumn('iris', 'new_col3', 'a/abs(-2)', datasources = opals)
str(iris)
# this will fail:
dssDeriveColumn('iris', 'new_col', 'Sepal.Length[1]', datasources = opals)
# as will this
dssDeriveColumn('iris', 'new_col', 'rnorm(length(iris$Sepal.Length), mean = mean(iris$Sepal.Length),1)', datasources = opals)

}
