% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssDim.R
\name{dssDim}
\alias{dssDim}
\title{Dimension of an object}
\usage{
dssDim(x = NULL, async = TRUE, wait = TRUE, datasources = NULL)
}
\arguments{
\item{x}{a character name of the data frame or matrix}

\item{async}{a logical, see datashield.aggregate}

\item{wait}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\value{
a list with one element for each node (and one $global element if type='combine'). Each element contains the dimensions of the object on the respective node.
}
\description{
Execute base::dim on the remote nodes
}
