% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssSetOption.R
\name{dssGetOption}
\alias{dssGetOption}
\title{Get options from remote R sessions}
\usage{
dssGetOption(opt = NULL, async = TRUE, datasources = NULL)
}
\arguments{
\item{opt}{the name of the option. If nULL all the session options will be retrieved}

\item{async}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\value{
a list with the option values
}
\description{
Get options from remote R sessions
}
\examples{
dssGetOption( datasources = my_datasources)
}
