% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssGM.R
\name{dssGM}
\alias{dssGM}
\title{Call mixtools on each source,}
\usage{
dssGM(what, cols = NULL, K, async = TRUE, datasources = NULL)
}
\arguments{
\item{what:}{The data frame.}

\item{cols:}{Vector of column names for the numeric variables to consider.}

\item{K:}{the number of expected Gaussian components in the mix.}
}
\description{
Call `solve.mixtools` on all remote sources.
}
