% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssHmisc.R
\name{dssHmisc}
\alias{dssHmisc}
\title{Call selected functions from the package Hmisc}
\usage{
dssHmisc(
  func,
  ...,
  newobj = NULL,
  async = TRUE,
  wait = TRUE,
  datasources = NULL
)
}
\arguments{
\item{func}{a character, the name of the function to call}

\item{...}{further arguments to be passed to the function (see the documentation of the Hmisc package).}

\item{newobj}{a character name of the new object to be created on the nodes if this is an assign function, ignored otherwise.}

\item{async}{a logical, see datashield.aggregate}

\item{wait}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
At the moment only the function 'rcorrcens' is implemented. Due to implementation specifics the 'data' argument
for this function is mandatory (and is specified as a character, the name of a data frame) and the 'formula' argument must be set as a character (enclosed in quotes)
}
