% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssIsUnique.R
\name{dssIsUnique}
\alias{dssIsUnique}
\title{Check if a vector is unique (tipically a dataframe column)}
\usage{
dssIsUnique(what, datasources = NULL)
}
\arguments{
\item{what}{a character name of the vector (or dataframe column specified with '$' - 'df$col')}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
Check if a vector is unique (tipically a dataframe column)
}
