% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssLM.R
\name{dssLM}
\alias{dssLM}
\title{Aggregate linear models into a single one for the estimation of beta.}
\usage{
dssLM(
  what,
  dep_var,
  expl_vars = NULL,
  async = TRUE,
  datasources = NULL,
  type = "combine"
)
}
\arguments{
\item{what:}{dataframe name.}

\item{dep_var:}{[string] the column name for the dependent variable (y).}

\item{expl_vars:}{[vector of strings] the column names for the explanatory variables.}

\item{type:}{[string] if 'split', return a list of the individual models of each datasource, if 'combine' return the global model.
If TRUE, merge them into a single model (default).}
}
\value{
the MLE estimator for beta for the pooled individuals.
}
\description{
Takes as input multiple results from `linregDSS` to make a single linear model,
   as if individuals from the multiple sources were pooled together before modeling.
}
