% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssMakeFactors.R
\name{dssMakeFactors}
\alias{dssMakeFactors}
\title{Transform the columns of a data frame into factors}
\usage{
dssMakeFactors(
  df,
  value = NULL,
  to.replace = NULL,
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{df}{a character, name of the dataframe}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\value{
nothing - the data frame is now transormed in the remote sessions
}
\description{
It only affects the non numeric columns (text or date)
}
