% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssMean.R
\name{dssMean}
\alias{dssMean}
\title{Mean of a vector}
\usage{
dssMean(what, na.rm = TRUE, datasources = NULL)
}
\arguments{
\item{what}{name of the vector}

\item{na.rm}{logical, should missing values be removed?}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\value{
a list containing one value per node plus one element for the global mean
}
\description{
Same as function mean()
}
