% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssMergeColumns.R
\name{dssMergeColumns}
\alias{dssMergeColumns}
\title{Merge columns in dataframe}
\usage{
dssMergeColumns(df, dictionary, datasources = NULL, async = TRUE)
}
\arguments{
\item{dictionary}{a list of the form (output_col_1 = list(input_col_1,... input_col_n), .... output_col_n = list(...))
The input columns will be read in sequence and the existing NAs will be filled with values from the subsequent columns in the list.
The end result is the output column.}

\item{datasources}{same as in datashield.assign}

\item{async}{same as in datashield.assign}
}
\value{
It doesn't return anything, the dataframe in the server session will now contain the new output columns.
}
\description{
Combine the contents of similar fields with similar codes according to a dictionary.
}
