% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssNlme.R
\name{dssNlme_groupedData}
\alias{dssNlme_groupedData}
\title{Calls nlme::groupedData on the provided arguments}
\usage{
dssNlme_groupedData(newobj, ..., async = TRUE, wait = TRUE, datasources = NULL)
}
\arguments{
\item{newobj}{a character, name of the new dataframe}

\item{...}{further arguments to be passed to nlme:groupedData. The data argument must pe passed as a character, the name of the input dataframe}

\item{async}{same as in datashield.assign}

\item{wait}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\description{
See the documentation of groupedData in package nlme
}
