% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssRange.R
\name{dssRange}
\alias{dssRange}
\title{Get the combined range of a (number of) vector(s) or dataframes}
\usage{
dssRange(..., type = "combine", datasources = NULL)
}
\arguments{
\item{...}{characters vector or dataframe names (must be present on all the nodes specified)}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\value{
combined range over the specified datasources
}
\description{
If one of the parameters is a dataframe, the ranges of all the numeric columns are returned (no need to subset for them)
}
