% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssRanger.R
\name{dssRanger}
\alias{dssRanger}
\title{Call selected functions from the package ranger}
\usage{
dssRanger(func, newobj, async, datasources, ...)
}
\arguments{
\item{func}{a character, the name of the function to call}

\item{newobj}{a character name of the new object to be created on the nodes.}

\item{async}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}

\item{...}{further arguments to be passed to the function (see the documentation of the VIM package). The names of objects (dataframes, vectors)
should be provided as a characters.}
}
\description{
At the moment only the functions 'ranger' and 'predict'  are implemented.
}
