% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssRbind.R
\name{dssRbind}
\alias{dssRbind}
\title{Rbind 2 dataframes}
\usage{
dssRbind(
  newobj,
  x,
  y,
  new.colnames = NULL,
  async = TRUE,
  wait = TRUE,
  datasources = NULL
)
}
\arguments{
\item{x, y}{characters names of the input dataframes.}

\item{new.colnames}{a vector the column names of the new dataframe (optional)}

\item{async}{same as in datashield.assign}

\item{wait}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}

\item{new.object}{a character name of the new dataframe}
}
\description{
Rbind 2 dataframes
}
