% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssRemoveOutliers.R
\name{dssRemoveOutliers}
\alias{dssRemoveOutliers}
\title{Remove outliers from all or some of the columns of a data frame and replace them with NAs.}
\usage{
dssRemoveOutliers(
  x,
  collist = NULL,
  sigmas = NULL,
  minval = -Inf,
  maxval = Inf,
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{x}{a character name of the data frame}

\item{collist}{a vector of columns. They must exist in x and be numeric.
If null, all the *numeric* columns from x will be used.}

\item{sigmas}{an integer, how many standard deviations will be kept on either side of the mean?
If NULL, minval and maxval will be taken into account.}

\item{minval}{either a number or a vector of numbers of length equal to collist (or to the number of numeric columns in x if the latter is null).
It sets the lower cutoff, values smaller than this will be replaced with NAs. Ignored if sigmas is set.}

\item{maxval}{either a number or a vector of numbers of length equal to collist (or to the number of numeric columns in x if the latter is null).
It sets the higher cutoff, values larger than this will be replaced with NAs. Ignored if sigmas is set.}

\item{async}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
Remove outliers from all or some of the columns of a data frame and replace them with NAs.
}
