% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssScale.R
\name{dssScale}
\alias{dssScale}
\title{Scale a dataframe}
\usage{
dssScale(
  symbol,
  what,
  center = TRUE,
  scale = TRUE,
  async = TRUE,
  wait = TRUE,
  datasources = NULL
)
}
\arguments{
\item{symbol}{a character name of the new, scaled dataframe}

\item{what}{a character name of the source dataframe}

\item{center}{either a logical value or a numeric vector of length equal to the number of columns of "what"}

\item{scale}{either a logical value or a numeric vector of length equal to the number of columns of "what"}

\item{async}{a logical, see datashield.aggregate}

\item{wait}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
execute base::scale() on the nodes
}
