% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssSetOption.R
\name{dssSetOption}
\alias{dssSetOption}
\title{Set  options in remote R sessions}
\usage{
dssSetOption(option_list, async = TRUE, wait = TRUE, datasources = NULL)
}
\arguments{
\item{option_list}{a list describing the options to be set}

\item{async}{a logical, see datashield.aggregate}

\item{wait}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\value{
TRUE if succesful
}
\description{
Set  options in remote R sessions
}
\examples{
dssSetOption(list(stringsAsFactors = TRUE, digits = 2), datasources = my_datasources)
}
