% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssShowFactors.R
\name{dssShowFactors}
\alias{dssShowFactors}
\title{Describe factors in a dataframe}
\usage{
dssShowFactors(what, show.all = FALSE, async = TRUE, datasources = NULL)
}
\arguments{
\item{what}{a character name of the dataframe}

\item{show.all}{a logical, default FALSE. If FALSE it limits to the first 50 levels, otherwise it returns all of them.}

\item{async}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
Show the levels for all the factors in a remote dataframe
}
