% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssSNF.R
\name{dssSNF}
\alias{dssSNF}
\title{Call selected functions from the package SNFtool}
\usage{
dssSNF(
  func,
  symbols,
  ...,
  action = c("keep", "return"),
  keep.name = NULL,
  async = TRUE,
  wait = TRUE,
  datasources = NULL
)
}
\arguments{
\item{func}{a character, the name of the function to call}

\item{symbols}{a character vector, name(s) of the input object(s) to the function.}

\item{...}{furhter arguments which will be passed to the function (see documentation of the SDFtool package)}

\item{action}{a character vector, what to do with the result. 'return' doesn't work for standardNormalization as it would disclose individual data.}

\item{keep.name}{a character, if action is 'keep', the name of the  output object on the remote nodes. If NULL, a name will be generated.}

\item{async}{a logical, see datashield.aggregate}

\item{wait}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
At the moment the functions 'standardNormalization','dist2','affinityMatrix' and 'spectralClustering' are implemented
}
\details{
The input to the function standardNormalizaton must be a dataframe containing the SUBJID colum and must *not* be transposed
}
