% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssPivot.R
\name{dssSuggestPivot}
\alias{dssSuggestPivot}
\title{Suggest a formula for widening a dataframe}
\usage{
dssSuggestPivot(
  what,
  cols = NULL,
  by.col = NULL,
  async = TRUE,
  wait = TRUE,
  datasources = NULL
)
}
\arguments{
\item{what}{a character name of the dataframe}

\item{cols}{a vector of columns used in the formula. They must exist in 'what'. If null, all the columns from 'what' will be used.}

\item{by.col}{typically the patient id, the column that will end up as key of the resulting 'wide' data.frame.}

\item{async}{a logical, see datashield.aggregate}

\item{wait}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
Makes a crude attempt at guessing an appropriate widening formula to feed into dssPivot.
The method is pretty basic, it should serve more as a starting point for further investigation
}
