% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{dssSwapKeys}
\alias{dssSwapKeys}
\title{Swap list keys}
\usage{
dssSwapKeys(l)
}
\arguments{
\item{l}{a list}
}
\value{
a list with swapped keys and values
}
\description{
Swap the first level of names of a list with the second level (second level can be more named lists or a simple vector).
Values are merged if necessary (see example)
}
\examples{
x <- list(a = list(b = c(1,2)), y = list(b = c(2,3)))
dssSwapKeys(x)
}
