% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssSynthpop.R
\name{dssSynthpop}
\alias{dssSynthpop}
\title{Generate synthetic data using the synthpop package}
\usage{
dssSynthpop(func, ..., async = TRUE, wait = TRUE, datasources = NULL)
}
\arguments{
\item{func}{a character, the name of the function to call (all functions are available).}

\item{...}{further arguments to be passed to the function (see the documentation of the Synthpop package).
Object names should be passed as character strings.}

\item{async}{a logical, see datashield.aggregate}

\item{wait}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
Generate synthetic data using the synthpop package
}
