% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsT.R
\name{dssT}
\alias{dssT}
\title{Transpose a  matrix on the remote nodes.}
\usage{
dssT(
  symbol,
  newobj = paste0(symbol, "_tr"),
  async = TRUE,
  wait = TRUE,
  datasources = NULL
)
}
\arguments{
\item{symbol}{a character, name of the object to transpose}

\item{newobj}{a character, name of the new, transposed object}

\item{async}{a logical, see datashield.aggregate}

\item{wait}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
See base::t()
}
