% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssTable.R
\name{dssTable}
\alias{dssTable}
\title{Execute the R table() function remotely. Counts under the privacy level threshold will be returned as 0.}
\usage{
dssTable(colname, type = "combine", async = TRUE, datasources = NULL)
}
\arguments{
\item{colname}{a character, the name of the column}

\item{type}{a character, should the results from different nodes be combined in one vector? Possibble values "split" or "combine"}

\item{async}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\value{
a table of counts
}
\description{
Execute the R table() function remotely. Counts under the privacy level threshold will be returned as 0.
}
