% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssUwot.R
\name{dssUwot}
\alias{dssUwot}
\title{Call selected functions from the package uwot}
\usage{
dssUwot(func, X, model = NULL, async = TRUE, datasources = NULL, ...)
}
\arguments{
\item{func}{a character, the name of the function to call}

\item{X, }{a character, the name of the dataframe containing the input data. The non-numeric columns are
automatically stripped of before running the function.}

\item{model}{a umap model, used only for the function umap_transform}

\item{async}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}

\item{...}{further arguments to be passed to the function (see the documentation of the uwot package).
The parameters ret_nn and ret_extra are not taken into consideration as they are potentially disclosive.}
}
\value{
a umap model (see the documentation for the respective functions)
}
\description{
Only the functions 'umap'  and 'umap_transform' are implemented
}
