% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssVar.R
\name{dssVar}
\alias{dssVar}
\title{Compute the variance of a given vector}
\usage{
dssVar(what, type = "combine", async = TRUE, datasources = NULL)
}
\arguments{
\item{what}{a character, the name of a numerical vector.}

\item{type}{a character which represents the type of analysis to carry out.
If type is set to 'combine', a global variance is calculated.
If type is set to 'split', the variance is calculated separately for each study.}

\item{async}{same as in datashield.assign}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\value{
A list containing:
In the case of 'combine', a list with a 'global' element (the global variance).
In the case of 'split', a list with as many elements as the 'datasources' argument (the respective variances).
}
\description{
Similar to the R function var, corrects the datashield function ds.var.
}
