% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssVIM.R
\name{dssVIM}
\alias{dssVIM}
\title{Call selected functions from the package VIM}
\usage{
dssVIM(func, newobj = NULL, async = TRUE, datasources = NULL, ...)
}
\arguments{
\item{func}{a character, the name of the function to call}

\item{newobj}{a character name of the new object to be created on the nodes if this is an assign function(i.e 'kNN'), ignored otherwise.
If an assign function is called without this, the original object will be modified in place.}

\item{async}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}

\item{...}{further arguments to be passed to the function (see the documentation of the VIM package). The first argument (typically a dataframe)
should be provided as a character.}
}
\description{
At the moment only the functions 'aggr' (which returns a result that can be visualised) and 'kNN' (which doesn't) are implemented
}
