% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssWaitForAsyncs.R
\name{dssWaitForAsyncs}
\alias{dssWaitForAsyncs}
\title{Wait for previous datashield commands}
\usage{
dssWaitForAsyncs(datasources, sleep = 0.5)
}
\arguments{
\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}

\item{sleep}{polling interval in seconds, default 0.5}
}
\value{
a list containing the results of the commands (if they were supposed to return anything) or raw(0)
}
\description{
It polls the current opal commands launched with async = TRUE
until all of them complete. Allows the launch of nonblocking datashield commands, useful in
web applications. The "sleep" argument must not be set too low (0.3 seconds is probably a good minimum)
otherwise it might look like a DOS attack and be blocked by the remote server.
}
