% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssGM.R
\docType{class}
\name{Gaussian-class}
\alias{Gaussian-class}
\alias{.Gaussian}
\title{Representation of a Gaussian component.}
\description{
Represents a set of normally-distributed points in R^n
    by their mean vector, covariance matrix and number of data points.
}

