% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssGM.R
\docType{class}
\name{Mixture-class}
\alias{Mixture-class}
\alias{.Mixture}
\title{Representation of a Gaussian Mixture.}
\description{
Just a wrapper on a list of Gaussian - to better understand what is in the list.
}

