% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{.encode.arg}
\alias{.encode.arg}
\title{Serialize function arguments for before sending them to the server}
\usage{
.encode.arg(some.object)
}
\arguments{
\item{some.object}{the object to be encoded}
}
\value{
encoded text with offending characters replaced by strings
}
\description{
Serialize to JSON, then encode base64 function arguments,
 then replace '+', '/' and '=' in the result in order to play nicely with the opal sentry
}
