% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssAddDaysInterval.R
\name{dssAddDaysInterval}
\alias{dssAddDaysInterval}
\title{Add and interval column to a dataframe}
\usage{
dssAddDaysInterval(
  newobj,
  df,
  description.list,
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{newobj}{a character, name of the new dataframe}

\item{df}{a character, name of the original dataframe}

\item{description.list}{a named list containing the name of the new colunn pointing to a list containing the start date and end date.
The format is list(<new column name> = list(start_date = '<existing date column>', end_date = '<existing date column>')) - see example.}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\description{
Add a column containing the difference between 2 date columns. The date columns must be in the same dataframe
}
