% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssCov.R
\name{dssCov}
\alias{dssCov}
\title{The covariance matrix of a data frame or matrix}
\usage{
dssCov(
  x,
  collist = NULL,
  type = "combine",
  na.rm = FALSE,
  wt = NULL,
  cor = FALSE,
  center = TRUE,
  method = "unbiased",
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{x}{a character name of the data frame}

\item{collist}{a vector of columns for the covariances. They must exist in x.
If null, all the *numeric* columns from x will be used.}

\item{type}{a character which represents the type of analysis to carry out.
If type is set to 'combine',  global column means are calculated if type is set to 'split', the column means are
calculated separately for each node.}

\item{na.rm}{a logical default FALSE, should NAs be removed before calculation? The TRUE value is processed only if type = 'split' and it calls the cov() function on each node with use = "pairwise.complete.obs"}

\item{wt}{if set, either a vector of weights or the name of such a vector. If the former, the same vector will be sent to
all the nodes. If the latter, different vectors can be first uploaded (with dssUpload) to the respective nodes. In both cases, the function will
execute cov.wt on the nodes and it will return a matrix per node irrespectve of the  the parameter "type" above}

\item{cor, }{center, method - parameters for the cov.wt function, ignored if wt is NULL (the default)}

\item{async}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\value{
a list with one element for each node (or one $global element if type='combine'). Each element contains
a vector with the respective column means, the covariance matrix and the number of rows
}
\description{
The covariance matrix of a data frame or matrix
}
