% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssRandomForest.R
\name{dssRandomForest}
\alias{dssRandomForest}
\title{Distributed random forests based on the randomForest package.}
\usage{
dssRandomForest(
  what,
  dep_var,
  expl_vars = NULL,
  testData = NULL,
  async = TRUE,
  wait = TRUE,
  datasources = NULL
)
}
\arguments{
\item{what:}{name of the training data frame on the server.}

\item{dep_var:}{[string] the response factor ("y"), i.e. the categories
that will be the leaves of each tree.}

\item{expl_vars:}{[vector[string]] the classification variables.}

\item{testData:}{[data frame] new data to classify using the forests.
It must have at least the columns in `expl_vars`.
(We want to predict the value of `dep_var` for it.)}
}
\value{
a randomForest object.

a list with members `$forests`: the individual forests from the nodes,
  and `$prediction`, the average prediction of `testData` (if given) by all nodes together.
}
\description{
Builds a random forest on each server node,
  as a model to predict classification of new data of the same type
  (new patients, same variables). The randomForest package needs to be
  installed on all nodes and on the client.
}
