% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssScale.R
\name{dssScale}
\alias{dssScale}
\title{Scale a dataframe}
\usage{
dssScale(
  symbol,
  what,
  center = TRUE,
  scale = TRUE,
  type = "combine",
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{symbol}{a character name of the new, scaled dataframe}

\item{what}{a character name of the source dataframe}

\item{center}{either a logical value or a numeric vector of length equal to the number of columns of "what"}

\item{scale}{either a logical value or a numeric vector of length equal to the number of columns of "what"}

\item{type}{a character, 'split' or 'combine'. If either center or scaled are set to  TRUE it defines the
scope for the means used for centering and/or standard deviations used for scaling. If 'combine', global values,
calculated accross all nodes are used, in case of 'split' the node local values are used. Default 'combine'.}

\item{async}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
execute base::scale() on the nodes
}
