% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssSynthpop.R
\name{dssSynthpop}
\alias{dssSynthpop}
\title{Generate synthetic data using the synthpop package}
\usage{
dssSynthpop(data, ..., async = TRUE, datasources = NULL)
}
\arguments{
\item{data}{a character, name of the data frame to use as base for the synthetic data}

\item{...}{further arguments to be passed to the function (see the documentation of the synthpop package).}

\item{async}{a logical, see datashield.aggregate}

\item{datasources}{a list of opal objects obtained after logging into the opal servers (see datashield.login)}
}
\description{
The only function available is synthpop::syn
}
